product and category fixes

+++ src/main/java/hu/bme/ecommercebackend/service/CategoryService.java
@@ -3,8 +3,10 @@ package hu.bme.ecommercebackend.service;
 import hu.bme.ecommercebackend.dto.Category.CategoryCreateDto;
 import hu.bme.ecommercebackend.dto.Category.CategoryDetailedDto;
 import hu.bme.ecommercebackend.dto.Category.CategoryDto;
+import hu.bme.ecommercebackend.dto.Common.ActionResponseDto;
 import hu.bme.ecommercebackend.model.Category;
 import hu.bme.ecommercebackend.repository.CategoryRepository;
+import hu.bme.ecommercebackend.repository.ProductRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
@@ -15,9 +17,11 @@ import java.util.stream.Collectors;
 @Service
 public class CategoryService {
     private final CategoryRepository categoryRepository;
+    private final ProductRepository productRepository;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, ProductRepository productRepository) {
         this.categoryRepository = categoryRepository;
+        this.productRepository = productRepository;
     }
 
     public CategoryDto getCategoryDtoById(Long id) {
@@ -40,6 +44,9 @@ public class CategoryService {
 
     public CategoryDto createCategoryDto(CategoryCreateDto newCategory) {
         Category categoryEntity = createCategory(newCategory);
+        categoryEntity.getSubCategories().forEach(category -> {
+            category.setParentCategory(categoryEntity);
+        });
         Category tempCategory = categoryRepository.save(categoryEntity);
         return new CategoryDto(tempCategory);
     }
@@ -68,8 +75,18 @@ public class CategoryService {
         return new CategoryDto(categoryEntity);
     }
 
-    public Boolean deleteCategoryWithId(Long id) {
-        categoryRepository.deleteById(id);
-        return true;
+    public ActionResponseDto deleteCategoryWithId(Long id) {
+        try {
+            boolean isInUse = this.productRepository.existsByCategoryId(id);
+            if(!isInUse) {
+                categoryRepository.deleteById(id);
+                return new ActionResponseDto(true,"");
+            } else {
+                return new ActionResponseDto(false,"A kategória használatban van, így törlése nem lehetséges.");
+            }
+
+        } catch (RuntimeException exception) {
+            return new ActionResponseDto(false,exception.getMessage());
+        }
     }
 }
